Track pending outbound transfers to keep funds unavailable
While we are in an `await ledger.transfer(...)`, we don't know if the
`ledger.account_balance` will reflect that. It is always safe for us to
think our available balance is lower than it is, so we will conservatively
withhold those funds until we are guaranteed the transfer has completed
and is reflected in the `ledger.account_balance`.